remove extension of PriceExtension in project-base
- the logic is moved to the parent class
